+++ pkg/pgmodel/metrics/ingest.go
@@ -5,6 +5,8 @@
 package metrics
 
 import (
+	"os"
+
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/timescale/promscale/pkg/util"
 )
@@ -62,24 +64,14 @@ var (
 			Help:      "Capacity of the ingest channel.",
 		}, []string{"type", "subsystem", "kind"},
 	)
-	IngestorChannelLen = prometheus.NewHistogramVec(
-		prometheus.HistogramOpts{
-			Namespace: util.PromNamespace,
-			Subsystem: "ingest",
-			Name:      "channel_len",
-			Help:      "Length of the ingestor channel.",
-			Buckets:   util.HistogramBucketsSaturating(0, 2, MetricBatcherChannelCap),
-		}, []string{"type", "subsystem", "kind"},
-	)
-	SampleCopierChannelLengthFunc func() float64
-	IngestorChannelLenCopier      = prometheus.NewGaugeFunc(
+	IngestorChannelLenBatcher = prometheus.NewGauge(
 		prometheus.GaugeOpts{
 			Namespace:   util.PromNamespace,
 			Subsystem:   "ingest",
 			Name:        "channel_len",
 			Help:        "Length of the ingestor channel.",
-			ConstLabels: map[string]string{"type": "metric", "subsystem": "copier", "kind": "sample"},
-		}, SampleCopierChannelLengthFunc,
+			ConstLabels: map[string]string{"type": "metric", "subsystem": "metric_batcher", "kind": "sample"},
+		},
 	)
 	IngestorFlushSeries = prometheus.NewHistogramVec(
 		prometheus.HistogramOpts{
@@ -187,7 +179,7 @@ func init() {
 		IngestorDecompressEarliest,
 		IngestorMaxSentTimestamp,
 		IngestorChannelCap,
-		IngestorChannelLen,
+		IngestorChannelLenBatcher,
 		IngestorFlushSeries,
 		IngestorInsertablesIngested,
 		IngestorInsertsPerBatch,
@@ -200,4 +192,27 @@ func init() {
 		IngestorRequests,
 		InsertBatchSize,
 	)
+}
+
+// RegisterCopierChannelLenMetric creates and registers the copier channel len metric with a callback
+// that should return the length of the channel.
+//
+// Note: ingestorChannelLenCopier metric depends on prometheus call to /metrics hence we need to update with
+// a callback. This is an odd one out from the other metrics in the ingestor as other metrics
+// are async to prometheus calls.
+func RegisterCopierChannelLenMetric(updater func() float64) {
+	r := prometheus.DefaultRegisterer
+	if val := os.Getenv("IS_TEST"); val == "true" {
+		r = prometheus.NewRegistry()
+	}
+	ingestorChannelLenCopier := prometheus.NewGaugeFunc(
+		prometheus.GaugeOpts{
+			Namespace:   util.PromNamespace,
+			Subsystem:   "ingest",
+			Name:        "channel_len",
+			Help:        "Length of the ingestor channel.",
+			ConstLabels: map[string]string{"type": "metric", "subsystem": "copier", "kind": "sample"},
+		}, updater,
+	)
+	r.MustRegister(ingestorChannelLenCopier)
 }